requirements.txt
flask
requests
pyyaml
pytest
reportlab
matplotlib
